+++ internal/api/handlers/bulletin_dto.go
@@ -1,11 +1,12 @@
+// Package handlers provides HTTP request handlers for all API endpoints.
 package handlers
 
 import "time"
 
 // BulletinResponse is the typed response for bulletin details
 type BulletinResponse struct {
-	ID          int       `json:"id"`
-	StationID   int       `json:"station_id"`
+	ID          int64     `json:"id"`
+	StationID   int64     `json:"station_id"`
 	StationName string    `json:"station_name"`
 	AudioURL    string    `json:"audio_url"`
 	Filename    string    `json:"filename"`
@@ -17,27 +18,27 @@ type BulletinResponse struct {
 
 // StationRef is a minimal station reference
 type StationRef struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
 
 // StoryRef is a minimal story reference
 type StoryRef struct {
-	ID    int    `json:"id"`
+	ID    int64  `json:"id"`
 	Title string `json:"title"`
 }
 
 // BulletinRef is a minimal bulletin reference
 type BulletinRef struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Filename string `json:"filename"`
 }
 
 // BulletinStoryResponse is the typed response for bulletin-story relationships
 type BulletinStoryResponse struct {
-	ID         int         `json:"id"`
-	BulletinID int         `json:"bulletin_id"`
-	StoryID    int         `json:"story_id"`
+	ID         int64       `json:"id"`
+	BulletinID int64       `json:"bulletin_id"`
+	StoryID    int64       `json:"story_id"`
 	StoryOrder int         `json:"story_order"`
 	CreatedAt  time.Time   `json:"created_at"`
 	Station    StationRef  `json:"station"`
