+++ internal/repository/base.go
@@ -1,3 +1,4 @@
+// Package repository provides data access abstractions for the Babbel application.
 package repository
 
 import (
@@ -43,7 +44,7 @@ func (r *BaseRepository[T]) TableName() string {
 }
 
 // GetByID retrieves a record by its ID.
-func (r *BaseRepository[T]) GetByID(ctx context.Context, id int) (*T, error) {
+func (r *BaseRepository[T]) GetByID(ctx context.Context, id int64) (*T, error) {
 	q := r.getQueryable(ctx)
 
 	var result T
@@ -60,7 +61,7 @@ func (r *BaseRepository[T]) GetByID(ctx context.Context, id int) (*T, error) {
 }
 
 // Exists checks if a record with the given ID exists.
-func (r *BaseRepository[T]) Exists(ctx context.Context, id int) (bool, error) {
+func (r *BaseRepository[T]) Exists(ctx context.Context, id int64) (bool, error) {
 	q := r.getQueryable(ctx)
 
 	var exists bool
@@ -88,7 +89,7 @@ func (r *BaseRepository[T]) Count(ctx context.Context) (int64, error) {
 }
 
 // Delete performs a hard delete of a record by ID.
-func (r *BaseRepository[T]) Delete(ctx context.Context, id int) error {
+func (r *BaseRepository[T]) Delete(ctx context.Context, id int64) error {
 	q := r.getQueryable(ctx)
 
 	query := fmt.Sprintf("DELETE FROM %s WHERE id = ?", r.tableName)
